Chapter 4
All done with chapter 4. Slightly different than chapter 3 but not by much, or at least enough for me to notice. was really more of just make a list, try this thing you can do, neat!

Did note down a cool concept about pointers that one of the examples brought up with me trying to do it the quick way and it not working though so that was exciting.