Update Notes.py
I was reading though the book last night and decided to move onto  the stuff I know I am missing or the bigger picture things.

I also noticed I have not been working on it as much as I would like. I have been doing lots more boot.dev and so need to find a better balance and be sure to not leave this behind.